requirements.txt
streamlit
openai
google-generativeai
pydub
python-dotenv